Extract cancellation logic into its own file and add test
